Made the level loop
The level now seamlessly loops, here's how:
- I use a camera to get the whole level
- I then render that to a render texture
- Those render textures are then placed either side of the level
- Once the player comes close to the edges, then their sprite is cloned (to cover up weirdness happening at the edges)
- After the player has travelled far enough outside the level, the player is then teleported to the same position in the actual level

Another way to do this would be to make the entire scene a prefab and then copy it either side, then add the teleports. This approach means that it's slightly easier to edit the level as we go along, and it may be better on performance, but that's not the biggest deal since it's a 2D side scroller.