Changed PowerUps to be more generic
Added a more Generic PowerUp script, in an attempt to once again make things more readable, BasicPowerUp is the main boy, attach that script to any thing that should act like a PowerUp and it will do it's thing

This also makes it so that the Player scripts only need to subscribe to a few PowerUp event objects, the Coffee powerup has some special behaviour though so it's likely that will have to stay for now until I perfect the solution.

I'm not an expert on events so if this breaks let me know!

I'll work on making our input implementation a bit better now, to allow for button remapping, allowing controllers and remembering user preferences.